clutch-covid-mdm-portal/requirements.txt
Cython==0.29.19
Flask==1.1.2
mysqlclient==1.4.6
Flask-SQLAlchemy==2.4.4
Flask-AppBuilder==3.1.1
PyMySQL==0.9.3
gunicorn==20.0.4
pyodbc==4.0.30
pymssql==2.1.4
distro==1.5.0
Flask-Admin==1.5.6
uWSGI==2.0.17.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
six==1.13.0
Werkzeug==0.16.0
pandas==1.2.1
scikit-learn==0.22
matplotlib==3.1.2
requests==2.22.0
SQLAlchemy==1.3.22
sasl==0.2.1
thrift==0.13.0
thrift-sasl==0.3.0
PyHive==0.6.1
hdfs==2.5.8
psycopg2-binary==2.8.5
Authlib==0.15.2
marshmallow==3.10.0
PyJWT==1.7.1
hl7==0.4.1
pysftp==0.2.9
APScheduler==3.7.0
Flask-Mail==0.9.1
Flask-Bootstrap==3.3.7.1
ERAlchemy==1.2.10
pysftp==0.2.9


